% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_diffs.r
\name{diff_1D_Ct}
\alias{diff_1D_Ct}
\title{Mean concentration values of one-dimensional diffusion during each time steps.}
\usage{
diff_1D_Ct(Lx, Tt, nt, nx, D, C_ini, C_lim)
}
\arguments{
\item{Lx}{Total length of slab in x direction, usually in mm.}

\item{Tt}{Total diffusion time, usually in second.}

\item{nt}{Number of time discretization.}

\item{nx}{Number of dimension x discretization}

\item{D}{Coefficient of diffusion. Must be constant value.}

\item{C_ini}{Initial concentration value inside the slab.}

\item{C_lim}{Boundary condition, written as a vector with one or two elements. If there is only one element, the other side of the slab will be presented as having Neumann boundary condition with flux = 0. If there are two elements, the first element is the dirichlet concentration on the left side, while the second element is the dirichlet concentration on the right side.}
}
\value{
A vector with \strong{nt} number of elements with the values of mean concentration in the slab for each T\emph{i}.
}
\description{
A wrapper of \emph{diff_1D} function,
returning the mean values of concentration during each
time step of the diffusion.
}
\examples{
Lx <- 5 #Length of slab, in mm
Tt <- 20 #Total measured diffusion time, in seconds
nt <- 100 #Number of time discretization, wherein the Tt will be divided into 100 equal parts (0, 0.2, 0.4, ..., 20)
nx <- 50 #Number of dimension discretization
D <- 0.05 #Coefficient of diffusion in mm^2/s.
C_ini <- 0.05 #Initial concentration inside the slab.
C_lim <- c(0.10,0.25) #Dirichlet boundary concentration diffusing into the slab
meanCt <- diff_1D_Ct(Lx, Tt, nt, nx, D, C_ini, C_lim)
}
