% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_interps.R
\name{vectInterp}
\alias{vectInterp}
\title{Interpolating the values between two vectors.}
\usage{
vectInterp(vect1, vect2, dist)
}
\arguments{
\item{vect1}{First vector variable.}

\item{vect2}{Second vector variable.}

\item{dist}{Desired interpolation value of distance between the two vectors.}
}
\value{
A vector with interpolated values between the first and the second vectors.
}
\description{
<<<<<<< HEAD
A wrapper of \emph{diff_1D} function,
=======
A wrapper of \emph{diff.1D} function,
>>>>>>> master
returning the mean values of concentration during each
time step of the diffusion.
}
\examples{
library(Rdiffsolver)
vect1 <- seq(1,100,length.out = 10)
vect2 <- seq(2.5,250,length.out = 10)
dist <- 0.12
vect_int <- vectInterp(vect1,vect2,dist)
}
