% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdiffsolver.R
\docType{package}
\name{Rdiffsolver}
\alias{Rdiffsolver}
\alias{Rdiffsolver-package}
\title{Rdiffsolver: A simple package for solving diffusion equation.}
\description{
The \strong{Rdiffsolver} package provides currently one basic function:
diff.1D, which can solve the diffusion equation in unidimensional
direction, with constant boundary conditions
(dirichlet and neumann with zero flux).
}
\details{
\strong{Methods:} Finite differences with backward Euler
implicit method (\emph{diff.1D} function).
}
\author{
\strong{Maintainer:} Ahmad Alkadri (email: \href{mailto:ahmad.alkadri@outlook.com}{ahmad.alkadri@outlook.com})
}
